requirements.txt
Django==1.11.9
django-cors-headers==2.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
pytz==2018.3
